server/project/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
datasets==3.5.0
deap==1.4.2
deprecation==2.1.0
dill==0.3.8
evaluate==0.4.3
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
gotrue==2.12.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.6.1
iniconfig==2.1.0
jax==0.4.30
jaxlib==0.4.30
jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
ml-dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt-einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.11.3
pydantic-core==2.33.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
torch-geometric==2.6.1
tqdm==4.67.1
transformers==4.51.2
triton==3.2.0
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
websockets==14.2
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
